API  API API!

+++ api/player.py
@@ -1,73 +1,57 @@
 from flask import Blueprint, request, jsonify
-from flask_restful import Api, Resource # used for REST API building
+from flask_restful import Api, Resource
 
 from model.players import Player
 
-# Change variable name and API name and prefix
-player_api = Blueprint('player_api', __name__,
-                   url_prefix='/api/players')
-
-# API docs https://flask-restful.readthedocs.io/en/latest/api.html
+player_api = Blueprint('player_api', __name__, url_prefix='/api/players')
 api = Api(player_api)
 
-class PlayerAPI:     
+class PlayerAPI:
     class Action(Resource):
         def post(self):
-            ''' Read data for json body '''
             body = request.get_json()
+            user = body.get('user')
             
-            ''' Avoid garbage in, error checking '''
-            # validate name
-            name = body.get('name')
-            if name is None or len(name) < 2:
-                return {'message': f'Name is missing, or is less than 2 characters'}, 210
-            # validate uid
-            uid = body.get('uid')
-            if uid is None or len(uid) < 2:
-                return {'message': f'User ID is missing, or is less than 2 characters'}, 210
-            # look for password and tokens
-            password = body.get('password')
-            tokens = body.get('tokens')
+            if user is None or len(user) < 2:
+                return {'message': f'Name is missing, or is less than 2 characters'}, 400
 
-            ''' #1: Key code block, setup PLAYER OBJECT '''
-            po = Player(name=name, 
-                        uid=uid,
-                        tokens=tokens)
-            
-            ''' Additional garbage error checking '''
-            # set password if provided
-            if password is not None:
-                po.set_password(password)            
-            
-            ''' #2: Key Code block to add user to database '''
-            # create player in database
+            po = Player(user=user, score=0)  # Set an initial score, change as needed
+                    
             player = po.create()
-            # success returns json of player
             if player:
-                return jsonify(player.read())
-            # failure returns error
-            return {'message': f'Processed {name}, either a format error or User ID {uid} is duplicate'}, 210
+                return jsonify(player.read()), 201
+            return {'message': f'Error creating player'}, 500
 
         def get(self):
-            players = Player.query.all()    # read/extract all players from database
-            json_ready = [player.read() for player in players]  # prepare output in json
-            return jsonify(json_ready)  # jsonify creates Flask response object, more specific to APIs than json.dumps
+            players = Player.query.all()
+            json_ready = [player.read() for player in players]
+            return jsonify(json_ready)
 
         def put(self):
-            body = request.get_json() # get the body of the request
-            uid = body.get('uid') # get the UID (Know what to reference)
+            body = request.get_json()
+            user = body.get('user')
             data = body.get('data')
-            player = Player.query.get(uid) # get the player (using the uid in this case)
-            player.update(data)
-            return f"{player.read()} Updated"
+            
+            if user is None:
+                return {'message': 'User is missing in the request data'}, 400
+            
+            player = Player.query.filter_by(user=user).first()
+            if player:
+                player.update(data)
+                return f"{player.read()} Updated", 200
+            return {'message': f'Player with user {user} not found'}, 404
 
         def delete(self):
             body = request.get_json()
-            uid = body.get('uid')
-            player = Player.query.get(uid)
-            player.delete()
-            return f"{player.read()} Has been deleted"
+            user = body.get('user')
+            
+            if user is None:
+                return {'message': 'User is missing in the request data'}, 400
+            
+            player = Player.query.filter_by(user=user).first()
+            if player:
+                player.delete()
+                return f"{player.read()} Has been deleted", 200
+            return {'message': f'Player with user {user} not found'}, 404
 
-
-    # building RESTapi endpoint, method distinguishes action
     api.add_resource(Action, '/')
